Refactor AuthWrapper & improve RenderRoutes code
