--- app/client/src/services/authService.js
@@ -1,9 +1,0 @@
-import axios from 'axios';
-const baseUrl = '/api/login/success';
-
-const loginSuccess = async (credentials) => {
-  const response = await axios.get(baseUrl, credentials);
-  return response.data;
-};
-
-export default loginSuccess;
