--- app/client/src/services/loginService.js
@@ -1,9 +1,0 @@
-import axios from 'axios';
-const baseUrl = '/api/login';
-
-const login = async (credentials) => {
-  const response = await axios.post(baseUrl, credentials);
-  return response.data;
-};
-// eslint-disable-next-line
-export default { login };
